{{ cookiecutter.project_slug }}/requirements-dev.txt
-r requirements.txt
backcall==0.1.0
decorator==4.4.0
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
coverage==4.5.3
pylint==2.3.1
pylint-flask==0.6
black==19.3b0
isort==4.3.21
pytest==5.1.1
pytest-cov==2.6.1
pytest-sugar==0.9.2
